Move util.
