+++ src/NetworkTest/testQuality/helpers/subscriberMOS.ts
@@ -3,7 +3,7 @@ import calculateThroughput from './calculateThroughput';
 import MOSState from './MOSState';
 import { OT } from '../../types/opentok';
 import { AV } from '../types/stats';
-import { getOr, last, nth } from '../../../util';
+import { getOr, last, nth } from '../../util';
 
 export type StatsListener = (error?: OT.OTError, stats?: OT.SubscriberStats) => void;
